requirements.txt
python-telegram-bot[webhooks]==21.7
python-dotenv==1.0.1
aiohttp==3.11.10
